// src/style/typography.d.ts
export declare const typography: {
    display: import("styled-components").RuleSet<object>;
    headline: import("styled-components").RuleSet<object>;
    title_02: import("styled-components").RuleSet<object>;
    title_01: import("styled-components").RuleSet<object>;
    subtitle_03: import("styled-components").RuleSet<object>;
    subtitle_02: import("styled-components").RuleSet<object>;
    subtitle_01: import("styled-components").RuleSet<object>;
    body_02: import("styled-components").RuleSet<object>;
    body_long_02: import("styled-components").RuleSet<object>;
    body_01: import("styled-components").RuleSet<object>;
    body_long_01: import("styled-components").RuleSet<object>;
    caption: import("styled-components").RuleSet<object>;
    caption_long: import("styled-components").RuleSet<object>;
};